condition: add GraphStructureCondition
A GraphStructureCondition is a pseudo-Condition that enables the user to
make defined changes to the scheduler graph using the same interface as
old (basic) Conditions. This allows for the easier definition of some
execution patterns that are difficult to do with basic Conditions by
leveraging the Scheduler's default consideration pattern based on the
graph structure.

A node has exactly one basic Condition, and may additionally have any
number of GraphStructureConditions that are applied in sequence.